Post-merge reconciliation: all packages build clean
Monorepo-wide fixes after merging 7 parallel branches:

Codegen + contracts
- shared/api/{schemas,events,errors}: z.number() -> z.number().int() on all
  count/duration fields so Go codegen emits int (not float64)
- shared/api/codegen.mjs: int64 -> int; canonicalize schemas before hashing
  (strip $schema/description/title, sort keys) so nested shapes dedup to
  top-level structs and we don't get duplicate ProgressSyncRequestState et al
- shared/voice/types.ts: regenerated with 120 keys including path's additions

control-plane
- internal/apitypes/helpers.go: Lang/Department type aliases, OKResponse,
  error code constants (ErrBadRequest et al), ValidLang/ValidDepartment
- internal/api: OK -> Ok, PtyWSURL -> PTYWSURL, respond.go accepts raw string
  for error code to decouple from ApiErrorCode typed enum
- go.sum regenerated via go mod tidy

Mobile
- apps/mobile/package.json: added uuid, react-native-svg,
  @react-native-async-storage/async-storage deps
- apps/mobile/lib/i18n.ts: re-export t, Lang, VoiceKey from shared-voice so
  convo components resolve their imports
- apps/mobile/lib/api.ts: added named exports requestMagicLink, claimMagicLink,
  patchMe, setDepartment, apiBaseUrl that path screens import
- apps/mobile/components/transcript/ToolCallCard.tsx: cast dynamic key to VoiceKey
- apps/mobile/app/role-pick.tsx: use haptics.confirm() wrapper instead of
  direct expo-haptics (keeps discipline of 3-signal vocabulary)

Web
- apps/web/tailwind.config.js: point at @learn/shared-tokens/tailwind-preset
  (proof's local shim deleted)
- apps/web/tailwind/tokens-preset.js: removed shim

Legacy cleanup
- web/: removed. MVP static Next.js from phase-0 was reference only

Verification
- pnpm typecheck: 10/10 packages pass
- go build ./... in control-plane, proxy, vm-image/claude-wrap, guest-agent
- go test ./internal/lesson/... pass
- go test proxy/tests pass

+++ shared/api/errors.ts
@@ -18,8 +18,8 @@ export type ApiErrorCode = z.infer<typeof ApiErrorCode>;
 export const ApiError = z.object({
   code: ApiErrorCode,
   message: z.string(),
-  retry_after_seconds: z.number().optional(),
-  queue_position: z.number().optional(),
+  retry_after_seconds: z.number().int().optional(),
+  queue_position: z.number().int().optional(),
   request_id: z.string(),
 });
 export type ApiError = z.infer<typeof ApiError>;
